+++ frontend/src/types.tsx
@@ -1,0 +1,11 @@
+export interface Currencies {
+  isoCode: string;
+  name: string;
+}
+
+export interface Currency {
+  isoCode: string;
+  name: string;
+  rate: number;
+  updatedAt: string;
+}
